refactor code in Typescript
