+++ src/index.ts
@@ -1,3 +1,2 @@
-// @flow
 import FlexView from './FlexView';
 export default FlexView;
